requirements.txt
# Ayende CX - Python Requirements
# Compatible with Python 3.10, 3.11, and 3.12

# Core Framework
Django>=5.0,<5.1

# API Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0

# CORS handling
django-cors-headers>=4.3.0

# Admin UI
django-unfold>=0.14.0

# Static files
whitenoise>=6.6.0

# Production server (for deployment later)
gunicorn>=21.2.0

# Database (PostgreSQL for production)
psycopg2-binary>=2.9.9

# Image handling
Pillow>=10.0.0

# Environment variables
python-dotenv>=1.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
django-debug-toolbar>=4.2.0

# Database URL parser (for production)
dj-database-url>=2.1.0